feat: allow operation according project role

+++ app/api/v1/endpoints/environments.py
@@ -20,6 +20,11 @@ from ....services.environment import (
     get_environment_by_id,
     update_environment,
 )
+from ....services.project_members import (
+    can_access_environment,
+    can_update_environment,
+    is_project_admin,
+)
 
 router = APIRouter(tags=['environments'])
 
@@ -31,6 +36,11 @@ async def create_new_environment(
     current_user: UserInDB = Depends(
         get_current_user)
 ):
+    has_admin_permission = await is_project_admin(db, environment.project_id, current_user.id)
+    if not has_admin_permission:
+        raise HTTPException(
+            status_code=401, detail="Not have permissions to create environment")
+
     new_environment = await create_environment(db, environment)
     if not new_environment:
         raise HTTPException(
@@ -45,6 +55,10 @@ async def get_environment(
     current_user: UserInDB = Depends(
         get_current_user)
 ):
+    can_read = await can_access_environment(db, id, current_user.id)
+    if not can_read:
+        raise HTTPException(status_code=401, detail='User is not authorized')
+
     environment = await get_environment_by_id(db, id)
     if not environment:
         raise HTTPException(
@@ -60,6 +74,11 @@ async def update_environment_route(
     current_user: UserInDB = Depends(
         get_current_user)
 ):
+    can_update = await can_update_environment(db, environment.project_id, current_user.id)
+    if not can_update:
+        raise HTTPException(
+            status_code=401, detail="Not have permissions to update environment")
+
     updated_environment = await update_environment(db, id, environment)
     if not updated_environment:
         raise HTTPException(
@@ -74,6 +93,13 @@ async def delete_environment_route(
     current_user: UserInDB = Depends(
         get_current_user)
 ):
+
+    environment = await get_environment_by_id(db, id)
+    has_admin_permission = await is_project_admin(db, environment.project_id, current_user.id)
+    if not has_admin_permission:
+        raise HTTPException(
+            status_code=401, detail="Not have permissions to create environment")
+
     if not await delete_environment(db, id):
         raise HTTPException(
             status_code=404, detail=f"Environment with id '{id}' not found")
